dcd-master/requirements.txt
scipy==1.5.2
pandas==1.2.2
ananke-causal==0.1.5